Fixes sign error in flux divergence of biharmonic stress and diffusive flux
Resolves #1702 noticed by @mukund-gupta .

In a previous PR we changed the notation of the closure term to indicate that it's the divergence of a flux, rather than specifically the divergence of a _viscous_ or _diffusive_ flux. This notational change required us to change the term's sign. However, we did not change the sign of the flux divergence for `AnisotropicBiharmonicDiffusivity`; we only changed the _name_ of the term.

This PR fixes that.