A method fixing (?) tracking tracer fields
To preface, the issue this fixes may be my miss understanding of how to use
tracked fields. Essentially the issue seems to be that to define an tracked
field you have to already have the model defined, but todo that you have to
have the particles defined. A work around for this was to make the model,
then define the particles, then redefine the model. This would run and not
error, but would result in the tracked field property of the particles not
actually being updated. Running it yet again appears to fix it and make it
work.

In order to fix this I've modified the `update_particle_properties!`
function to explicitly use the models tracer fields when it is a tracer
field being tracked. I've also updated the tests which seem to run.