Added BiogeochemicalForcing infrastructure (see note)
We can't setup Continuous forcing with the normal functions because
it makes interpolation operators dependent on the index of tracers
so the full model needs to be setup before you can create the forcing
which doesn't work here.

I guess the normal continuous forcing is as complicated as it is for
a reason so we might have todo something else at some point, maybe
for GPU compatability?