Switched to a Latitude-Longitude Grid
In this commit, we switched the
(a) grid from RectilinearGrid to LatitudeLongitudeGrid;
(b) momentum_advection from WENO() to VectorInvariant(); and
(c) coriolis from BetaPlane(latitude = 45) to HydrostaticSphericalCoriolis().

Unfortunately, the numerical solution is no longer stable and blows up after a
few time steps. Coarsening the spatial resolution and reducing the time step (to
lower the Courant number) has no effect.