Created tracer advection animation on cubed sphere
(a) Renamed the script multi_region_cubedsphere.jl to
    multi_region_cubed_sphere.jl;
(b) Wrapped the testing of the heatsphere! and heatlatlon! plotting
    functions of multi_region_cubed_sphere.jl into a function;
(c) Imported the function heatsphere! from multi_region_cubed_sphere.jl to
    multi_region_tracer_advection.jl to create an animation of the tracer
    advection on the cubed sphere;
(d) Wrapped the testing of the multi region tracer advection of
    multi_region_tracer_advection.jl into a function;
(e) Since the velocities are still zero, the tracer does not advect yet.