Added option to fill streamfunction halos manually
Added options to:
(a) fill out the streamfunction halos manually;
(b) plot the streamfunction and velocities on the interior and halo nodes.

If I fill out the streamfunction values manually on the halos, they all turn out
to be zero. I verified that by printing out (and plotting) the streamfunction
values on the interior and the halo nodes. After a bit of investigation, I
realized that λnodes and φnodes are not initialized at locations ffa. They hold
zero values, and so does the streamfunction from the analytical formula.