Fixed minor bugs and cleaned up code
The Rossby-Haurwitz simulation runs indefinitely without encountering stability
issues. Our immediate goal is to perform a detailed comparison between the grid
metrics and velocities in the halos of this version (0.82.0) and the current
version of Oceananigans. To facilitate this, the code segment responsible for
executing the simulation has been temporarily commented out.