Added simplified initial velocity for debugging
This update introduces simplified linear options for the initial velocity field,
aimed at enhancing the debugging process for halos. The added configurations are
as follows:

(a) u = c₁ y and v = 0 with c₁ = 1, so that ζ = -c₁ = -1 = constant;
(b) u = c₁ y and v = c₂ x with c₁ = 1 and c₂ = 2, so that ζ = c₂ - c₁ = 2 - 1
    = 1 = constant.

These options are specifically designed to simplify the analysis of the initial
vorticity field, ζ,  within both the interior and halo regions of each panel on
the cubed sphere, with a view to streamlining the debugging process. However,
the current implementation of these options has been temporarily commented out.