Enforce total buoyancy flux BC in tilted geometry example
This example includes no explicit modification to the BCs on buoyancy,
meaning that it defaults to a no-flux BC (or `FluxBoundaryCondition()`)
on buoyancy.

Following equation (6) of Wenegrat and Thomas (2020), we instead enforce
a no-normal diffusive flux boundary condition on the *total* buoyancy, i.e.
the perturbation plus the constantly-stratified `BackgroundField`. Because
Oceananigans.jl does not allow diffusion to act on `BackgroundField`s, the
background part of the no-flux BC is missing. However, we here can enforce it
by specifying a perturbation flux BC that matches the implied background flux.

Other minor changes:
- Added equally-spaced buoyancy surfaces to movie panels
- Guess timestep by minimum of advective and diffusive timescales,
instead of just the advective timescale