Enforce total buoyancy flux BC in tilted geometry example (#3581)
* Enforce total buoyancy flux BC in tilted geometry example

This example includes no explicit modification to the BCs on buoyancy,
meaning that it defaults to a no-flux BC (or `FluxBoundaryCondition()`)
on buoyancy.

Following equation (6) of Wenegrat and Thomas (2020), we instead enforce
a no-normal diffusive flux boundary condition on the *total* buoyancy, i.e.
the perturbation plus the constantly-stratified `BackgroundField`. Because
Oceananigans.jl does not allow diffusion to act on `BackgroundField`s, the
background part of the no-flux BC is missing. However, we here can enforce it
by specifying a perturbation flux BC that matches the implied background flux.

Other minor changes:
- Added equally-spaced buoyancy surfaces to movie panels
- Guess timestep by minimum of advective and diffusive timescales,
instead of just the advective timescale

* Update examples/tilted_bottom_boundary_layer.jl

Co-authored-by: Gregory L. Wagner <gregory.leclaire.wagner@gmail.com>

* Imeplement Greg's suggestions

* Add equations to explain the choice of bottom BC on buoyancy

* More intentional phrasing to explain perturbation BC

Co-authored-by: Gregory L. Wagner <gregory.leclaire.wagner@gmail.com>

* More transparent assignment of bottom BC

Co-authored-by: Gregory L. Wagner <gregory.leclaire.wagner@gmail.com>

* Rename background field to match mathematical notation

Co-authored-by: Gregory L. Wagner <gregory.leclaire.wagner@gmail.com>

* Rename background field to match mathematical notation, follow-up

Co-authored-by: Gregory L. Wagner <gregory.leclaire.wagner@gmail.com>

---------

Co-authored-by: Henri Drake <henrifdrake@ch3212.ps.uci.edu>
Co-authored-by: Gregory L. Wagner <gregory.leclaire.wagner@gmail.com>
Co-authored-by: Gregory L. Wagner <wagner.greg@gmail.com>