Add extrema(c::AbstractField)
Currently, `extrema(field)` fails with scalar indexing on CUDA. The
reason for this is that `extrema(c::AbstractField)` is not defined,
so we fall back to `extrema(a::AbstractArray; dims, kw...)`.

This commit adds `extrema(field)` by calling `minimum(field)` and
`maximum(field)` separately. Note that there's an opportunity for
optimization here, where both the extrama are found in the same
reduction.